Default ImGui D3D12 to use shader binaries
These changes are intended to simplify that path on whether ImGui should
use shader source or shader binaries...and make things a little
friendlier for Wine.

- Changed USE_DXIIVK_WORK_AROUND to USE_SHADER_BINARIES.
- Removed GGP platform note in imgui_impl_dx12.cpp.
- Defaulted path for ImGui D3D12 to use shader binaries instead of
  compiling from source.
- Fixed casting warning.

+++ include/ppx/grfx/dx/d3dcompile_util.h
@@ -31,7 +31,7 @@ public:
             itr               = fileNameToContents.insert({pFileName, fs::load_file(filePath)}).first;
         }
         *ppData = itr->second.data();
-        *pBytes = itr->second.size();
+        *pBytes = static_cast<UINT>(itr->second.size());
         return S_OK;
     }
